refactor: minor changes
-Change button component name to linkbutton since they are links. -Reword header description. -Make minor adjustments to other components.
