+++ src/components/UI/NavBar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "./Button";
+import Button from "./LinkButton";
 import Styles from "./NavBar.module.css";
 import MobileNav from "./MobileNav";
 import NavBtnData from "../../data/NavBtnData";
